third_party/blink/web_tests/external/wpt/wasm/jsapi/proto-from-ctor-realm-expected.txt

This is a testharness.js-based test.
Found 75 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] WebAssembly.Module: cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `-1` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `""` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `"str"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: cross-realm NewTarget with `symbol "Symbol()"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: bound cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: bound bound cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: bound Proxy of cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: Proxy of cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: Proxy of Proxy of cross-realm NewTarget with `-0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Module: Proxy of bound cross-realm NewTarget with `NaN` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `-1` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `""` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `"str"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: cross-realm NewTarget with `symbol "Symbol()"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: bound cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: bound bound cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: bound Proxy of cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: Proxy of cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: Proxy of Proxy of cross-realm NewTarget with `-0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Instance: Proxy of bound cross-realm NewTarget with `NaN` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `-1` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `""` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `"str"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: cross-realm NewTarget with `symbol "Symbol()"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: bound cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: bound bound cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: bound Proxy of cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: Proxy of cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: Proxy of Proxy of cross-realm NewTarget with `-0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Memory: Proxy of bound cross-realm NewTarget with `NaN` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `-1` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `""` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `"str"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: cross-realm NewTarget with `symbol "Symbol()"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: bound cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: bound bound cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: bound Proxy of cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: Proxy of cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: Proxy of Proxy of cross-realm NewTarget with `-0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Table: Proxy of bound cross-realm NewTarget with `NaN` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `-1` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `""` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `"str"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: cross-realm NewTarget with `symbol "Symbol()"` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: bound cross-realm NewTarget with `undefined` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: bound bound cross-realm NewTarget with `null` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: bound Proxy of cross-realm NewTarget with `false` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: Proxy of cross-realm NewTarget with `true` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: Proxy of Proxy of cross-realm NewTarget with `-0` prototype
  assert_true: expected true got false
[FAIL] WebAssembly.Global: Proxy of bound cross-realm NewTarget with `NaN` prototype
  assert_true: expected true got false
Harness: the test ran to completion.
